requirements.txt
python-telegram-bot==20.7
firebase-admin==6.3.0
pydub==0.25.1
faster-whisper==0.10.0
ffmpeg-python==0.2.0
